+++ lib/src/core/library.dart
@@ -1,10 +1,10 @@
-import "dart:math" as math;
-
 import "package:fling_units/fling_units.dart" as f;
 
 export "../measurements/speed.dart";
 
 export "dimension.dart";
+
+export "precision/precision.dart";
 
 part "extension_num.dart";
 part "unit.dart";
@@ -15,7 +15,6 @@ part "measurement_prefix.dart";
 part "measurement.dart";
 part "measurement2.dart";
 part "measurement3.dart";
-part "precision.dart";
 part "measurement2_builder.dart";
 
 part "../measurements/volume.dart";